Add Model Method, Clarify Field Names, Rename Project
This commit changed the winter_hiatus field by making it a method. Now the field will check the current month against the months during which there is a hiatus from street sweeping and return a boolean.

There was a possibility that just having "even" and "odd" would not be clear to users.  So, these were changed to "even_side" and "odd_side" respectively.

Finally, this project is primarily about street sweeping, so the project name was changed to reflect this.